refactor: split mcp/handlers.rs 3223 LOC → 7 sub-modules (GREEN)
Pure refactor — zero behavior change. Split monolith into:
- mod.rs (172 LOC): dispatch router + shared helpers
- channel.rs (76 LOC): reply, react, edit_message, download_attachment
- comms.rs (573 LOC): send_to/delegate/report/broadcast/inbox
- instance.rs (644 LOC): create/delete/start/describe/replace/interrupt
- task.rs (82 LOC): task board + decisions + teams
- schedule.rs (30 LOC): schedules + deployments
- ci.rs (121 LOC): watch_ci/unwatch_ci/checkout_repo/release_repo
- tests.rs (1579 LOC): all original tests preserved

Invariant test enforces <700 LOC per handler file.
§3.5.11 test-first: 17e16e8 (RED) fails; this commit passes.
