+++ src/mcp/handlers/tests.rs
@@ -1,0 +1,1579 @@
+use super::*;
+use serde_json::json;
+
+fn tmp_home(name: &str) -> std::path::PathBuf {
+    use std::sync::atomic::{AtomicU32, Ordering};
+    static COUNTER: AtomicU32 = AtomicU32::new(0);
+    let id = COUNTER.fetch_add(1, Ordering::Relaxed);
+    let dir = std::env::temp_dir().join(format!(
+        "agend-handlers-test-{}-{}-{}",
+        std::process::id(),
+        name,
+        id
+    ));
+    std::fs::create_dir_all(&dir).ok();
+    dir
+}
+
+// `validate_branch` tests live in `src/agent_ops.rs` — migrated there
+// as part of Task #9 Option C.
+
+// merge_metadata tests
+#[test]
+fn merge_metadata_no_file() {
+    let home = tmp_home("merge_meta_no_file");
+    let mut info = json!({"name": "agent1", "state": "idle"});
+    merge_metadata(&home, "agent1", &mut info);
+    // Should not crash, info unchanged
+    assert_eq!(info["name"], "agent1");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn merge_metadata_merges_fields() {
+    let home = tmp_home("merge_meta_fields");
+    let meta_dir = home.join("metadata");
+    std::fs::create_dir_all(&meta_dir).ok();
+    std::fs::write(
+        meta_dir.join("agent1.json"),
+        r#"{"display_name": "Dev Agent", "custom": 42}"#,
+    )
+    .ok();
+    let mut info = json!({"name": "agent1", "state": "idle"});
+    merge_metadata(&home, "agent1", &mut info);
+    assert_eq!(info["display_name"], "Dev Agent");
+    assert_eq!(info["custom"], 42);
+    assert_eq!(info["name"], "agent1"); // original preserved
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// save_metadata tests
+#[test]
+fn save_and_load_metadata() {
+    let home = tmp_home("save_meta");
+    save_metadata(&home, "agent1", "display_name", json!("My Agent"));
+    save_metadata(&home, "agent1", "version", json!(2));
+    let content = std::fs::read_to_string(home.join("metadata/agent1.json")).expect("read");
+    let meta: serde_json::Value = serde_json::from_str(&content).expect("parse");
+    assert_eq!(meta["display_name"], "My Agent");
+    assert_eq!(meta["version"], 2);
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn save_metadata_creates_dir() {
+    let home = tmp_home("save_meta_dir");
+    assert!(!home.join("metadata").exists());
+    save_metadata(&home, "agent1", "key", json!("value"));
+    assert!(home.join("metadata").exists());
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// get_submit_key tests
+use crate::agent_ops::get_submit_key;
+
+#[test]
+fn get_submit_key_default() {
+    let home = tmp_home("submit_key");
+    // No fleet.yaml → default \r
+    let key = get_submit_key(&home, "agent1");
+    assert_eq!(key, "\r");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn get_submit_key_from_fleet() {
+    let home = tmp_home("submit_key_fleet");
+    let yaml = r#"defaults:
+  backend: claude
+instances:
+  dev:
+    role: "Developer"
+"#;
+    std::fs::write(home.join("fleet.yaml"), yaml).ok();
+    let key = get_submit_key(&home, "dev");
+    // Claude Code preset submit_key is "\r" or similar
+    assert!(!key.is_empty());
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// --- cleanup_working_dir ---
+
+#[test]
+fn cleanup_agend_workspace_removes_entire_dir() {
+    let home = tmp_home("cleanup_ws");
+    let ws = home.join("workspace").join("test-agent");
+    std::fs::create_dir_all(&ws).ok();
+    std::fs::write(ws.join("somefile.txt"), "data").ok();
+    std::fs::write(ws.join("opencode.json"), "{}").ok();
+
+    cleanup_working_dir(&home, "test-agent", &ws);
+    assert!(!ws.exists(), "workspace dir should be fully removed");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn cleanup_user_dir_only_removes_agend_files() {
+    let home = tmp_home("cleanup_user");
+    let user_dir = tmp_home("cleanup_user_proj");
+
+    // Create user file + agend-generated files
+    std::fs::write(user_dir.join("main.rs"), "fn main() {}").ok();
+    std::fs::write(user_dir.join("opencode.json"), "{}").ok();
+    std::fs::write(user_dir.join("mcp-config.json"), "{}").ok();
+    std::fs::create_dir_all(user_dir.join(".claude")).ok();
+    std::fs::write(user_dir.join(".claude/settings.local.json"), "{}").ok();
+
+    cleanup_working_dir(&home, "agent1", &user_dir);
+
+    // User file preserved
+    assert!(user_dir.join("main.rs").exists(), "user file must survive");
+    // Agend files removed
+    assert!(!user_dir.join("opencode.json").exists());
+    assert!(!user_dir.join("mcp-config.json").exists());
+    assert!(!user_dir.join(".claude/settings.local.json").exists());
+
+    std::fs::remove_dir_all(&home).ok();
+    std::fs::remove_dir_all(&user_dir).ok();
+}
+
+#[test]
+fn cleanup_removes_metadata() {
+    let home = tmp_home("cleanup_meta");
+    let ws = home.join("workspace").join("agent1");
+    std::fs::create_dir_all(&ws).ok();
+
+    std::fs::create_dir_all(home.join("metadata")).ok();
+    std::fs::write(home.join("metadata/agent1.json"), "{}").ok();
+
+    cleanup_working_dir(&home, "agent1", &ws);
+
+    assert!(!home.join("metadata/agent1.json").exists());
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn cleanup_nonexistent_dir_no_panic() {
+    let home = tmp_home("cleanup_nodir");
+    let fake = std::path::PathBuf::from("/tmp/nonexistent-agend-test-dir");
+    // Should not panic
+    cleanup_working_dir(&home, "agent1", &fake);
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// ---------------------------------------------------------------------
+// FleetEvent emission tests (Stage B-UX PR-A, design §2)
+//
+// These tests share two pieces of global state: (1) `AGEND_HOME` env
+// var, read by `crate::home_dir()` inside `handle_tool`; (2) the
+// crate-wide `ux_sink_registry()` singleton. Both are process-scoped,
+// so the tests serialize through `fleet_test_guard()` and swap in a
+// `RecordingSink` per case via `clear_for_test`.
+//
+// Each positive test carries at least one pin (Reviewer Contract v0.1
+// §4) on the _source_ of the captured field — e.g. `task_id` must come
+// from the handler's `correlation_id` arg, `decision_id` must come
+// from `decisions::post`'s return, `recipients` must come from the
+// filtered `sent` vec — not from the caller's raw args.
+// ---------------------------------------------------------------------
+
+use crate::channel::sink_registry::registry as ux_sink_registry;
+use crate::channel::ux_event::{FleetEvent, UxEvent, UxEventSink};
+use parking_lot::{Mutex, MutexGuard};
+use std::sync::Arc;
+
+fn fleet_test_guard() -> MutexGuard<'static, ()> {
+    static GUARD: Mutex<()> = Mutex::new(());
+    GUARD.lock()
+}
+
+struct Recorder {
+    events: Mutex<Vec<UxEvent>>,
+}
+
+impl Recorder {
+    fn new() -> Arc<Self> {
+        Arc::new(Self {
+            events: Mutex::new(Vec::new()),
+        })
+    }
+
+    fn snapshot(&self) -> Vec<UxEvent> {
+        self.events.lock().clone()
+    }
+}
+
+impl UxEventSink for Recorder {
+    fn emit(&self, event: &UxEvent) {
+        self.events.lock().push(event.clone());
+    }
+}
+
+/// Set `AGEND_HOME` to a fresh temp dir with a minimal fleet.yaml
+/// (so `get_submit_key` fallbacks resolve), wipe the global sink
+/// registry, and register a fresh `Recorder`. Returns the recorder
+/// and the temp-home path so callers can clean up.
+fn setup_recorder(tag: &str) -> (Arc<Recorder>, std::path::PathBuf) {
+    let home = tmp_home(tag);
+    // Still set AGEND_HOME for sub-calls inside handle_tool_with_home
+    // that read home_dir() (e.g. get_submit_key fallback, inbox ops).
+    // Safe: fleet_test_guard serializes these tests, and the cross-module
+    // racers (instructions.rs, telegram.rs) no longer set AGEND_HOME.
+    std::env::set_var("AGEND_HOME", &home);
+    // Minimal fleet so send_to's `get_submit_key` lookup resolves
+    // (fallback path on unreachable daemon still needs submit_key).
+    let yaml = "defaults:\n  backend: claude\ninstances:\n  target:\n    role: Test\n  sender:\n    role: Test\n";
+    std::fs::write(home.join("fleet.yaml"), yaml).ok();
+    let rec = Recorder::new();
+    ux_sink_registry().clear_for_test();
+    ux_sink_registry().register(rec.clone() as Arc<dyn UxEventSink>);
+    (rec, home)
+}
+
+#[test]
+fn delegate_task_emits_fleet_event() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_delegate");
+
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "do the thing"}),
+        "sender",
+    );
+    // Must have succeeded (API-path or fallback); both populate "target".
+    assert!(
+        is_ok_result(&result),
+        "delegate_task should succeed: {result}"
+    );
+
+    let events = rec.snapshot();
+    assert_eq!(events.len(), 1, "expected one FleetEvent: {events:?}");
+    match &events[0] {
+        UxEvent::Fleet(FleetEvent::DelegateTask {
+            from,
+            to,
+            summary,
+            task_id,
+        }) => {
+            assert_eq!(from, "sender");
+            assert_eq!(to, "target");
+            assert_eq!(summary, "do the thing");
+            // Pin: `delegate_task` has no id slot; correlation surfaces
+            // later via `report_result.correlation_id`. Must be None.
+            assert!(task_id.is_none(), "task_id must be None for delegate");
+        }
+        other => panic!("expected DelegateTask, got {other:?}"),
+    }
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn report_result_emits_with_correlation_id() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_report");
+
+    // Pin: task_id must source from the `correlation_id` arg, not
+    // any other string field. Use a distinctive value so a stray
+    // field aliasing bug would fail the assert below.
+    let result = handle_tool(
+        "report_result",
+        &json!({
+            "target_instance": "target",
+            "summary": "done",
+            "correlation_id": "AGD-42",
+        }),
+        "sender",
+    );
+    assert!(
+        is_ok_result(&result),
+        "report_result should succeed: {result}"
+    );
+
+    let events = rec.snapshot();
+    assert_eq!(events.len(), 1);
+    match &events[0] {
+        UxEvent::Fleet(FleetEvent::ReportResult {
+            from,
+            to,
+            summary,
+            task_id,
+        }) => {
+            assert_eq!(from, "sender");
+            assert_eq!(to, "target");
+            assert_eq!(summary, "done");
+            assert_eq!(
+                task_id.as_deref(),
+                Some("AGD-42"),
+                "task_id must come from correlation_id arg"
+            );
+        }
+        other => panic!("expected ReportResult, got {other:?}"),
+    }
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn report_result_empty_correlation_id_maps_to_none() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_report_empty");
+
+    // Pin: empty `correlation_id` must collapse to None so the
+    // renderer omits the id rather than showing "()" — filter-empty
+    // is the specified normalization.
+    let _ = handle_tool(
+        "report_result",
+        &json!({
+            "target_instance": "target",
+            "summary": "done",
+            "correlation_id": "",
+        }),
+        "sender",
+    );
+
+    let events = rec.snapshot();
+    assert_eq!(events.len(), 1);
+    match &events[0] {
+        UxEvent::Fleet(FleetEvent::ReportResult { task_id, .. }) => {
+            assert!(
+                task_id.is_none(),
+                "empty correlation_id must normalize to None, got {task_id:?}"
+            );
+        }
+        other => panic!("expected ReportResult, got {other:?}"),
+    }
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn post_decision_with_sender_emits_fleet_event() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_decision");
+
+    let result = handle_tool(
+        "post_decision",
+        &json!({"title": "use X over Y", "content": "because Z"}),
+        "sender",
+    );
+    let posted_id = result["id"]
+        .as_str()
+        .unwrap_or_else(|| panic!("post_decision must return id: {result}"))
+        .to_string();
+
+    let events = rec.snapshot();
+    assert_eq!(events.len(), 1);
+    match &events[0] {
+        UxEvent::Fleet(FleetEvent::PostDecision {
+            by,
+            title,
+            decision_id,
+        }) => {
+            assert_eq!(by, "sender");
+            assert_eq!(title, "use X over Y");
+            // Pin: decision_id must come from `decisions::post`'s
+            // returned id (authoritative, nanosecond-stamped), NOT
+            // from any args the caller passed. Args have no id.
+            assert_eq!(
+                decision_id, &posted_id,
+                "decision_id must source from decisions::post result"
+            );
+        }
+        other => panic!("expected PostDecision, got {other:?}"),
+    }
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn post_decision_anonymous_does_not_emit_fleet_event() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_decision_anon");
+
+    // Ensure no AGEND_INSTANCE_NAME fallback kicks in for the handler.
+    std::env::remove_var("AGEND_INSTANCE_NAME");
+
+    // Anonymous call — `instance_name` empty, no env fallback.
+    // Decisions module itself must still succeed (anonymous contract),
+    // only fleet mirroring is suppressed. Pin: absence of emission.
+    let result = handle_tool(
+        "post_decision",
+        &json!({"title": "anon call", "content": "no author"}),
+        "",
+    );
+    assert!(
+        result["id"].as_str().is_some(),
+        "post_decision still succeeds anonymously: {result}"
+    );
+
+    let events = rec.snapshot();
+    assert!(
+        events.is_empty(),
+        "anonymous post_decision must NOT emit: {events:?}"
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn broadcast_emits_with_resolved_recipients() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_broadcast");
+
+    // `targets: ["target", "sender"]` — handler must self-filter
+    // `sender` out before computing the recipient set. Pin: the
+    // emitted `recipients` field comes from the filtered `sent`
+    // vec, NOT from the raw `args["targets"]`.
+    let result = handle_tool(
+        "broadcast",
+        &json!({
+            "message": "heads up",
+            "targets": ["target", "sender"],
+        }),
+        "sender",
+    );
+    // broadcast always returns {"sent_to": [...], "count": ...}
+    assert_eq!(result["count"].as_u64(), Some(1));
+
+    let events = rec.snapshot();
+    assert_eq!(events.len(), 1);
+    match &events[0] {
+        UxEvent::Fleet(FleetEvent::Broadcast {
+            from,
+            recipients,
+            summary,
+        }) => {
+            assert_eq!(from, "sender");
+            assert_eq!(summary, "heads up");
+            assert_eq!(
+                recipients,
+                &vec!["target".to_string()],
+                "recipients must be the self-filtered `sent` vec"
+            );
+        }
+        other => panic!("expected Broadcast, got {other:?}"),
+    }
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn broadcast_empty_targets_does_not_emit() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_broadcast_empty");
+
+    // `targets: ["sender"]` — all recipients are the sender itself,
+    // so the self-filter leaves `sent` empty. Pin: skip-emit on
+    // empty fan-out so fleet_binding isn't spammed with "a → *0".
+    let result = handle_tool(
+        "broadcast",
+        &json!({
+            "message": "alone",
+            "targets": ["sender"],
+        }),
+        "sender",
+    );
+    assert_eq!(result["count"].as_u64(), Some(0));
+
+    let events = rec.snapshot();
+    assert!(
+        events.is_empty(),
+        "empty-recipient broadcast must NOT emit: {events:?}"
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn send_to_instance_does_not_emit_fleet_event() {
+    // Negative pin (design §8 exclusion): routine `send_to_instance`
+    // is a point-to-point DM and intentionally NOT mirrored into
+    // fleet_binding. If a future refactor accidentally routes it
+    // through `FleetEvent`, this test fails.
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_send_to_excluded");
+
+    // send_to_instance takes `instance_name` (or `target`), not
+    // `target_instance` — use the real arg shape so the negative
+    // pin actually exercises the success path.
+    let result = handle_tool(
+        "send_to_instance",
+        &json!({"instance_name": "target", "message": "hi"}),
+        "sender",
+    );
+    assert!(
+        is_ok_result(&result),
+        "send_to_instance should succeed: {result}"
+    );
+
+    let events = rec.snapshot();
+    assert!(
+        events.is_empty(),
+        "send_to_instance must NOT emit FleetEvent (design §8 exclusion): {events:?}"
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn request_information_does_not_emit_fleet_event() {
+    // Negative pin (design §8 exclusion): `request_information` is
+    // a point-to-point query, not a fleet-visible coordination
+    // action. Guards against over-eager emission by future edits.
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_request_info_excluded");
+
+    let result = handle_tool(
+        "request_information",
+        &json!({"target_instance": "target", "question": "what is X?"}),
+        "sender",
+    );
+    // The send may fail in test env (no active daemon + env race on
+    // AGEND_HOME). That's fine — this test's purpose is the negative
+    // fleet-event assertion below, not send success. When it does
+    // succeed (inbox fallback), the result still must not carry an
+    // event. When it fails, the early-return path also must not emit.
+    let _ = result;
+
+    let events = rec.snapshot();
+    assert!(
+        events.is_empty(),
+        "request_information must NOT emit FleetEvent (design §8 exclusion): {events:?}"
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// ---------------------------------------------------------------------
+// S2d provenance injection tests (Stage B-UX PR-C, design §6 + §4 Q4)
+//
+// In test env, no Telegram bot is configured, so
+// `inject_provenance` hits `resolve_channel()` and bails with
+// `no_channel_configured`. That's exactly the failure mode these
+// pins exercise: we want the handler to stay clean when provenance
+// can't be delivered (main `send_to` result untouched, fleet event
+// still emitted), AND we want the `tracing::warn!` to actually fire
+// so operators have a signal that routing might be broken.
+// ---------------------------------------------------------------------
+
+/// Negative pin (main-path isolation): when `inject_provenance`
+/// fails, the handler's returned JSON must NOT carry any provenance
+/// text, and the FleetEvent::DelegateTask must STILL emit.
+///
+/// Why this pin: a naive refactor that threaded `inject_provenance`
+/// into `send_to`'s pipeline (rather than fanning it out as a
+/// side-channel) could pollute the caller's response or suppress
+/// the fleet event on provenance failure — this pin catches both.
+#[test]
+fn delegate_task_main_response_clean_when_provenance_fails() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("fleet_prov_main_clean");
+
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "do the thing"}),
+        "sender",
+    );
+
+    // Main path untouched: the handler still returns an ok result.
+    assert!(
+        is_ok_result(&result),
+        "delegate_task must succeed even when provenance fails: {result}"
+    );
+
+    // Main response clean: no provenance-failure bleed into the
+    // caller-visible JSON. Check both the rendered text and raw
+    // JSON so a future refactor that tucks the error into a nested
+    // field still trips the pin.
+    let rendered = result.to_string();
+    assert!(
+        !rendered.to_lowercase().contains("provenance"),
+        "main response leaked provenance text: {rendered}"
+    );
+    assert!(
+        !rendered.contains("⬅️"),
+        "main response leaked provenance tag glyph: {rendered}"
+    );
+
+    // Fleet visibility preserved: DelegateTask still reaches the sink.
+    let events = rec.snapshot();
+    assert_eq!(
+        events.len(),
+        1,
+        "FleetEvent must still emit when provenance fails: {events:?}"
+    );
+    assert!(
+        matches!(&events[0], UxEvent::Fleet(FleetEvent::DelegateTask { .. })),
+        "unexpected event variant: {:?}",
+        events[0]
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+/// Failure-visibility pin (DESIGN §4 Q4): `inject_provenance` failure
+/// MUST produce a `tracing::warn!` record, not a silent drop.
+///
+/// Why this pin: Q4 explicitly overrode the initial silent-log bias
+/// with warn-level because provenance failures can indicate real
+/// routing bugs (wrong topic_id for target) that otherwise decay
+/// silently. A future edit that downgrades to `debug!` or removes
+/// the `tracing::warn!` call entirely would lose that signal; this
+/// pin asserts the warn record is actually emitted.
+///
+/// `tracing-test`'s `#[traced_test]` attaches a capturing subscriber
+/// for the duration of this test; `logs_contain` scans captured
+/// records for a substring match.
+#[test]
+#[tracing_test::traced_test]
+fn delegate_task_provenance_failure_logs_tracing_warn() {
+    let _g = fleet_test_guard();
+    let (_rec, home) = setup_recorder("fleet_prov_warn");
+
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "do the thing"}),
+        "sender",
+    );
+    assert!(is_ok_result(&result), "handler must succeed: {result}");
+
+    // The warn carries the DESIGN §6 signature phrase. Pinning the
+    // exact message substring (not just "warn level fired") keeps
+    // the pin from passing on an unrelated warn that happened to
+    // fire during the handler run.
+    assert!(
+        logs_contain("S2d provenance injection failed"),
+        "DESIGN §4 Q4 warn record was not emitted",
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// -----------------------------------------------------------------
+// Track 1 PR-1 tests (design §7)
+// -----------------------------------------------------------------
+
+#[test]
+fn set_waiting_on_persists_and_clears() {
+    let _g = fleet_test_guard();
+    let (_rec, home) = setup_recorder("waiting_on_set");
+
+    // Set waiting_on
+    let result = handle_tool(
+        "set_waiting_on",
+        &json!({"condition": "review from at-dev-4"}),
+        "sender",
+    );
+    assert!(
+        is_ok_result(&result),
+        "set_waiting_on should succeed: {result}"
+    );
+    assert_eq!(result["waiting_on"], "review from at-dev-4");
+
+    // Value-source pin: return value `since` must match metadata file.
+    let returned_since = result["since"].as_str().expect("since in return");
+    let meta: Value = serde_json::from_str(
+        &std::fs::read_to_string(home.join("metadata/sender.json")).expect("read meta"),
+    )
+    .expect("parse meta");
+    assert_eq!(meta["waiting_on"], "review from at-dev-4");
+    assert_eq!(
+        meta["waiting_on_since"].as_str().expect("since in file"),
+        returned_since,
+        "return value since must match persisted timestamp (value-source pin)"
+    );
+
+    // Clear
+    let result = handle_tool("set_waiting_on", &json!({"condition": ""}), "sender");
+    assert_eq!(result["cleared"], true);
+    let meta: Value = serde_json::from_str(
+        &std::fs::read_to_string(home.join("metadata/sender.json")).expect("read meta"),
+    )
+    .expect("parse meta");
+    assert!(
+        meta["waiting_on"].is_null(),
+        "waiting_on must be null after clear"
+    );
+    assert!(
+        meta["waiting_on_since"].is_null(),
+        "waiting_on_since must be null after clear"
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn implicit_heartbeat_recorded_on_tool_call() {
+    let _g = fleet_test_guard();
+    let (_rec, home) = setup_recorder("heartbeat_rec");
+
+    // Any tool call should record heartbeat
+    let _ = handle_tool("inbox", &json!({}), "sender");
+
+    // Resolve meta_path from home_dir() *after* the tool call — on
+    // Windows CI, parallel tests can mutate AGEND_HOME between
+    // setup_recorder's set_var and handle_tool's home_dir() read.
+    // Using home_dir() here matches wherever handle_tool actually wrote.
+    let actual_home = crate::home_dir();
+    let meta_path = actual_home.join("metadata/sender.json");
+    let meta: Value =
+        serde_json::from_str(&std::fs::read_to_string(&meta_path).expect("read meta"))
+            .expect("parse meta — atomic write must produce valid JSON");
+    let hb = meta["last_heartbeat"]
+        .as_str()
+        .expect("last_heartbeat must be present after tool call");
+    // Must be a valid RFC3339 timestamp
+    assert!(
+        chrono::DateTime::parse_from_rfc3339(hb).is_ok(),
+        "last_heartbeat must be valid RFC3339: {hb}"
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+    if actual_home != home {
+        std::fs::remove_dir_all(&actual_home).ok();
+    }
+}
+
+#[test]
+fn waiting_on_exposed_via_merge_metadata() {
+    let _g = fleet_test_guard();
+    let (_rec, home) = setup_recorder("waiting_on_merge");
+
+    // Set waiting_on
+    let _ = handle_tool(
+        "set_waiting_on",
+        &json!({"condition": "delegation result"}),
+        "sender",
+    );
+
+    // Resolve home after tool call — parallel tests can shift AGEND_HOME
+    // on Windows CI (same class as PR #65).
+    let actual_home = crate::home_dir();
+
+    // Simulate what list_instances does: merge_metadata into agent info
+    let mut info = json!({"name": "sender", "agent_state": "thinking"});
+    merge_metadata(&actual_home, "sender", &mut info);
+    assert_eq!(
+        info["waiting_on"], "delegation result",
+        "merge_metadata must surface waiting_on"
+    );
+    assert!(
+        info["waiting_on_since"].as_str().is_some(),
+        "merge_metadata must surface waiting_on_since"
+    );
+    assert!(
+        info["last_heartbeat"].as_str().is_some(),
+        "merge_metadata must surface last_heartbeat"
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+    if actual_home != home {
+        std::fs::remove_dir_all(&actual_home).ok();
+    }
+}
+
+#[test]
+fn atomic_save_metadata_no_tmp_residue() {
+    let home = tmp_home("atomic_no_tmp");
+    save_metadata(&home, "agent1", "key", json!("value"));
+    let meta_dir = home.join("metadata");
+    let tmp_files: Vec<_> = std::fs::read_dir(&meta_dir)
+        .expect("read metadata dir")
+        .filter_map(|e| e.ok())
+        .filter(|e| e.path().extension().is_some_and(|ext| ext == "tmp"))
+        .collect();
+    assert!(
+        tmp_files.is_empty(),
+        "no .tmp residue after atomic write: {tmp_files:?}"
+    );
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn set_waiting_on_rejects_anonymous_caller() {
+    let _g = fleet_test_guard();
+    let (_rec, home) = setup_recorder("waiting_on_anon");
+    // Ensure no Sender resolves from env either
+    std::env::remove_var("AGEND_INSTANCE_NAME");
+    let result = handle_tool("set_waiting_on", &json!({"condition": "whatever"}), "");
+    assert!(
+        result["error"].is_string(),
+        "must err on anonymous caller: {result}"
+    );
+    // Must NOT have created metadata/.json
+    assert!(
+        !home.join("metadata/.json").exists(),
+        "no metadata written for anon caller"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// -----------------------------------------------------------------
+// PR #66 follow-up: F1 tests + F2 multi-message regression pin
+// -----------------------------------------------------------------
+
+#[test]
+fn metadata_persisted_on_pending_pickup() {
+    let _g = fleet_test_guard();
+    let (_rec, home) = setup_recorder("meta_pickup");
+
+    // Simulate what handle_message does: write pending_pickup_ids
+    save_metadata(
+        &home,
+        "sender",
+        "pending_pickup_ids",
+        json!([{"kind": "telegram", "msg_id": "42"}]),
+    );
+
+    let meta: Value = serde_json::from_str(
+        &std::fs::read_to_string(home.join("metadata/sender.json")).expect("read"),
+    )
+    .expect("parse");
+    let arr = meta["pending_pickup_ids"]
+        .as_array()
+        .expect("must be array");
+    assert_eq!(arr.len(), 1);
+    assert_eq!(arr[0]["kind"], "telegram");
+    assert_eq!(arr[0]["msg_id"], "42");
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn agent_picked_up_emitted_on_inbox_drain() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("pickup_emit");
+
+    // Seed pending_pickup_ids in metadata
+    save_metadata(
+        &home,
+        "sender",
+        "pending_pickup_ids",
+        json!([{"kind": "telegram", "msg_id": "99"}]),
+    );
+
+    // Seed an inbox message so drain is non-empty
+    let _ = crate::inbox::enqueue(
+        &home,
+        "sender",
+        crate::inbox::InboxMessage {
+            schema_version: 0,
+            id: None,
+            read_at: None,
+            thread_id: None,
+            parent_id: None,
+            task_id: None,
+            force_meta: None,
+            correlation_id: None,
+            reviewed_head: None,
+            from: "user:test".to_string(),
+            text: "hello".to_string(),
+            kind: Some("telegram".to_string()),
+            timestamp: chrono::Utc::now().to_rfc3339(),
+            channel: None,
+            delivery_mode: None,
+            attachments: vec![],
+            in_reply_to_msg_id: None,
+            in_reply_to_excerpt: None,
+        },
+    );
+
+    let _ = handle_tool("inbox", &json!({}), "sender");
+
+    let events = rec.snapshot();
+    let pickups: Vec<_> = events
+        .iter()
+        .filter(|e| matches!(e, UxEvent::AgentPickedUp { .. }))
+        .collect();
+    assert_eq!(pickups.len(), 1, "expected 1 AgentPickedUp: {events:?}");
+    if let UxEvent::AgentPickedUp { origin_msg, agent } = &pickups[0] {
+        assert_eq!(origin_msg.id, "99");
+        assert_eq!(agent, "sender");
+    }
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn agent_picked_up_fires_for_all_pending_messages() {
+    let _g = fleet_test_guard();
+    let (rec, home) = setup_recorder("pickup_multi");
+
+    // Seed 3 pending pickup IDs (simulating 3 rapid user messages)
+    save_metadata(
+        &home,
+        "sender",
+        "pending_pickup_ids",
+        json!([
+            {"kind": "telegram", "msg_id": "10"},
+            {"kind": "telegram", "msg_id": "11"},
+            {"kind": "telegram", "msg_id": "12"},
+        ]),
+    );
+
+    // Seed inbox message so drain is non-empty
+    let _ = crate::inbox::enqueue(
+        &home,
+        "sender",
+        crate::inbox::InboxMessage {
+            schema_version: 0,
+            id: None,
+            read_at: None,
+            thread_id: None,
+            parent_id: None,
+            task_id: None,
+            force_meta: None,
+            correlation_id: None,
+            reviewed_head: None,
+            from: "user:test".to_string(),
+            text: "burst".to_string(),
+            kind: Some("telegram".to_string()),
+            timestamp: chrono::Utc::now().to_rfc3339(),
+            channel: None,
+            delivery_mode: None,
+            attachments: vec![],
+            in_reply_to_msg_id: None,
+            in_reply_to_excerpt: None,
+        },
+    );
+
+    let _ = handle_tool("inbox", &json!({}), "sender");
+
+    let events = rec.snapshot();
+    let pickups: Vec<_> = events
+        .iter()
+        .filter(|e| matches!(e, UxEvent::AgentPickedUp { .. }))
+        .collect();
+    assert_eq!(
+        pickups.len(),
+        3,
+        "F2 pin: must emit AgentPickedUp for ALL pending messages, not just last: {events:?}"
+    );
+    // Verify IDs match
+    let ids: Vec<&str> = pickups
+        .iter()
+        .filter_map(|e| {
+            if let UxEvent::AgentPickedUp { origin_msg, .. } = e {
+                Some(origin_msg.id.as_str())
+            } else {
+                None
+            }
+        })
+        .collect();
+    assert_eq!(ids, vec!["10", "11", "12"]);
+
+    // Verify pending_pickup_ids cleared after drain
+    let meta: Value = serde_json::from_str(
+        &std::fs::read_to_string(home.join("metadata/sender.json")).expect("read"),
+    )
+    .expect("parse");
+    assert!(
+        meta["pending_pickup_ids"].is_null(),
+        "pending_pickup_ids must be cleared after drain"
+    );
+
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// --- Sprint 5: target validation + team routing ---
+
+#[test]
+fn test_send_to_nonexistent_target_returns_error_and_no_inbox() {
+    // F1+F2: daemon down + ghost target → error, NO inbox file created.
+    let _g = fleet_test_guard();
+    let home = tmp_home("send-nonexist");
+    std::env::set_var("AGEND_HOME", &home);
+    // No fleet.yaml → target doesn't exist anywhere.
+    let result = handle_tool(
+        "send_to_instance",
+        &json!({"instance_name": "ghost-agent", "message": "hello"}),
+        "sender",
+    );
+    assert!(
+        result.get("error").is_some(),
+        "send to nonexistent target must return error, got: {result}"
+    );
+    let ghost_inbox = home.join("inbox").join("ghost-agent.jsonl");
+    assert!(
+        !ghost_inbox.exists(),
+        "inbox file must NOT be created for nonexistent target"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn test_delegate_task_resolves_team_to_orchestrator_inbox() {
+    // F3: delegate_task to team name → resolved to orchestrator,
+    // verify the actual inbox recipient is the orchestrator.
+    let _g = fleet_test_guard();
+    let home = tmp_home("delegate-team");
+    // fleet.yaml for instance validation, teams.json for team resolution
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  dev-lead:\n    backend: claude\n  dev-impl:\n    backend: claude\n",
+    )
+    .ok();
+    // teams.json is the runtime store used by resolve_team_orchestrator
+    std::fs::write(
+            home.join("teams.json"),
+            r#"{"schema_version":1,"teams":[{"name":"dev","members":["dev-lead","dev-impl"],"orchestrator":"dev-lead","description":null,"created_at":"2026-01-01T00:00:00Z"}]}"#,
+        )
+        .ok();
+    std::env::set_var("AGEND_HOME", &home);
+
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "dev", "task": "test task"}),
+        "dev-impl",
+    );
+    // Should not error — team resolved to dev-lead.
+    let err = result["error"].as_str().unwrap_or("");
+    assert!(
+        !err.contains("not found"),
+        "delegate_task to team name should resolve, got error: {err}"
+    );
+    // Result should target dev-lead (orchestrator), not "dev" (team).
+    assert_eq!(
+        result["target"].as_str().unwrap_or(""),
+        "dev-lead",
+        "delegate_task must resolve team to orchestrator in result"
+    );
+    // No inbox for the team name itself.
+    let team_inbox = home.join("inbox").join("dev.jsonl");
+    assert!(
+        !team_inbox.exists(),
+        "inbox must NOT be created for team name 'dev'"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// --- Sprint 6: delivery_mode ---
+
+#[test]
+fn test_send_to_inbox_fallback_mode() {
+    // Daemon down → fallback path → delivery_mode = "inbox_fallback"
+    let _g = fleet_test_guard();
+    let home = tmp_home("delivery-fallback");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  receiver:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    let result = handle_tool(
+        "send_to_instance",
+        &json!({"instance_name": "receiver", "message": "test"}),
+        "sender",
+    );
+    assert_eq!(
+        result["delivery_mode"].as_str(),
+        Some("inbox_fallback"),
+        "daemon-down path must set delivery_mode=inbox_fallback: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn test_describe_message_shows_delivery_mode() {
+    // Verify describe_message returns delivery_mode when stored on the message.
+    let _g = fleet_test_guard();
+    let home = tmp_home("describe-dm");
+    std::env::set_var("AGEND_HOME", &home);
+    // Seed an inbox message with delivery_mode
+    let msg = crate::inbox::InboxMessage {
+        schema_version: 1,
+        id: Some("m-dm-test".into()),
+        from: "test".into(),
+        text: "hello".into(),
+        kind: None,
+        timestamp: chrono::Utc::now().to_rfc3339(),
+        channel: None,
+        read_at: Some(chrono::Utc::now().to_rfc3339()),
+        thread_id: None,
+        parent_id: None,
+        delivery_mode: Some("inbox_fallback".into()),
+        force_meta: None,
+        correlation_id: None,
+        reviewed_head: None,
+        task_id: None,
+        attachments: vec![],
+        in_reply_to_msg_id: None,
+        in_reply_to_excerpt: None,
+    };
+    let inbox_dir = home.join("inbox");
+    std::fs::create_dir_all(&inbox_dir).ok();
+    std::fs::write(
+        inbox_dir.join("agent1.jsonl"),
+        format!("{}\n", serde_json::to_string(&msg).unwrap()),
+    )
+    .ok();
+    let result = handle_tool(
+        "describe_message",
+        &json!({"message_id": "m-dm-test", "instance": "agent1"}),
+        "agent1",
+    );
+    assert_eq!(result["status"], "read");
+    assert_eq!(
+        result["delivery_mode"].as_str(),
+        Some("inbox_fallback"),
+        "describe_message must show delivery_mode: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// --- Sprint 8: busy gate + interrupt ---
+
+#[test]
+fn test_delegate_task_busy_returns_structured_response() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("busy-gate");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  target:\n    backend: claude\n  sender:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    // Create and claim a task for target
+    crate::tasks::handle(
+        &home,
+        "target",
+        &json!({"action": "create", "title": "busy work"}),
+    );
+    let tasks = crate::tasks::handle(&home, "target", &json!({"action": "list"}));
+    let tid = tasks["tasks"][0]["id"].as_str().unwrap();
+    crate::tasks::handle(&home, "target", &json!({"action": "claim", "id": tid}));
+
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "new work"}),
+        "sender",
+    );
+    assert_eq!(result["busy"], true, "must return busy: {result}");
+    assert!(
+        result["current_task"]["id"].is_string(),
+        "must have current_task.id: {result}"
+    );
+    assert!(result["options"].is_array(), "must have options: {result}");
+    assert!(
+        result["suggestion"].is_string(),
+        "must have suggestion: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn test_delegate_task_force_true_bypasses_busy_gate() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("interrupt-bypass");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  target:\n    backend: claude\n  sender:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    crate::tasks::handle(
+        &home,
+        "target",
+        &json!({"action": "create", "title": "busy"}),
+    );
+    let tasks = crate::tasks::handle(&home, "target", &json!({"action": "list"}));
+    let tid = tasks["tasks"][0]["id"].as_str().unwrap();
+    crate::tasks::handle(&home, "target", &json!({"action": "claim", "id": tid}));
+
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "urgent", "force": true, "force_reason": "critical bug"}),
+        "sender",
+    );
+    assert!(
+        result.get("busy").is_none(),
+        "interrupt=true must bypass busy gate: {result}"
+    );
+    assert!(
+        result.get("error").is_none() || !result["error"].as_str().unwrap_or("").contains("busy"),
+        "must not error on busy: {result}"
+    );
+    // Verify force_meta persisted in receiver's inbox
+    let msgs = crate::inbox::drain(&home, "target");
+    assert!(!msgs.is_empty(), "target must have inbox message");
+    let msg = &msgs[0];
+    assert!(
+        msg.force_meta.is_some(),
+        "force_meta must be set on inbox message: {:?}",
+        msg.force_meta
+    );
+    let meta = msg.force_meta.as_ref().unwrap();
+    assert!(meta.forced);
+    assert_eq!(meta.reason, "critical bug");
+    assert!(!meta.forced_at.is_empty());
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn test_delegate_task_force_true_without_reason_rejected() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("interrupt-no-reason");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  target:\n    backend: claude\n  sender:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    crate::tasks::handle(
+        &home,
+        "target",
+        &json!({"action": "create", "title": "busy"}),
+    );
+    let tasks = crate::tasks::handle(&home, "target", &json!({"action": "list"}));
+    let tid = tasks["tasks"][0]["id"].as_str().unwrap();
+    crate::tasks::handle(&home, "target", &json!({"action": "claim", "id": tid}));
+
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "urgent", "force": true}),
+        "sender",
+    );
+    assert!(
+        result["error"].as_str().unwrap_or("").contains("reason"),
+        "interrupt without reason must error: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn test_delegate_task_idle_target_normal_delivery() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("idle-target");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  target:\n    backend: claude\n  sender:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    // No claimed tasks for target
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "normal work"}),
+        "sender",
+    );
+    assert!(
+        result.get("busy").is_none(),
+        "idle target must not return busy: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// --- Sprint 9 Gap 5: second_reviewer flag ---
+
+#[test]
+fn test_delegate_task_second_reviewer_flag_requires_reason() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("sr-no-reason");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  target:\n    backend: claude\n  sender:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "review PR", "second_reviewer": true}),
+        "sender",
+    );
+    assert!(
+        result["error"]
+            .as_str()
+            .unwrap_or("")
+            .contains("second_reviewer_reason"),
+        "second_reviewer=true without reason must error: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn test_delegate_task_second_reviewer_with_reason_ok() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("sr-with-reason");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  target:\n    backend: claude\n  sender:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    let result = handle_tool(
+        "delegate_task",
+        &json!({
+            "target_instance": "target",
+            "task": "review PR",
+            "second_reviewer": true,
+            "second_reviewer_reason": "high-risk protocol change"
+        }),
+        "sender",
+    );
+    assert!(
+        result.get("error").is_none()
+            || !result["error"]
+                .as_str()
+                .unwrap_or("")
+                .contains("second_reviewer"),
+        "second_reviewer with reason must not error on flag: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn test_delegate_task_no_second_reviewer_flag_default_behavior() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("sr-default");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  target:\n    backend: claude\n  sender:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "normal work"}),
+        "sender",
+    );
+    // No second_reviewer flag → no error related to it
+    let err = result["error"].as_str().unwrap_or("");
+    assert!(
+        !err.contains("second_reviewer"),
+        "default (no flag) must not error on second_reviewer: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// ── interrupt tool tests ──
+
+#[test]
+fn test_interrupt_esc_params_contains_exact_esc_byte() {
+    let params = super::interrupt_esc_params("my-agent");
+    assert_eq!(params["method"], "inject");
+    assert_eq!(params["params"]["name"], "my-agent");
+    // Verify the data field is exactly the ESC byte (0x1b)
+    let data = params["params"]["data"]
+        .as_str()
+        .expect("data must be string");
+    assert_eq!(data.len(), 1, "ESC byte must be exactly 1 byte");
+    assert_eq!(data.as_bytes()[0], 0x1b, "data must be ESC byte (0x1b)");
+    assert_eq!(params["params"]["raw"], true, "must be raw inject");
+}
+
+#[test]
+fn test_interrupt_reason_header_format() {
+    let header = crate::inbox::format_event_header("interrupt", &[("reason", "priority task")]);
+    assert!(header.contains("[AGEND-MSG]"), "must have header prefix");
+    assert!(
+        header.contains("kind=interrupt"),
+        "must have interrupt kind"
+    );
+    assert!(
+        header.contains("reason=priority task"),
+        "must contain reason"
+    );
+    assert!(!header.contains('\n'), "must be single line");
+}
+
+#[test]
+fn test_interrupt_handler_validates_target() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("interrupt-validate");
+    std::env::set_var("AGEND_HOME", &home);
+
+    // Missing target
+    let r = handle_tool("interrupt", &json!({}), "caller");
+    assert!(r["error"].as_str().unwrap().contains("missing"));
+
+    // Invalid target name
+    let r = handle_tool("interrupt", &json!({"target": "../escape"}), "caller");
+    assert!(r.get("error").is_some());
+
+    // Valid target but no daemon → reaches inject path
+    let r = handle_tool("interrupt", &json!({"target": "valid-agent"}), "caller");
+    let err = r["error"].as_str().unwrap_or("");
+    assert!(
+        err.contains("not reachable") || err.contains("API unavailable"),
+        "valid target must reach inject path: {err}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// --- Sprint 10: backwards-compat for old interrupt/reason names ---
+
+#[test]
+fn test_delegate_task_old_interrupt_true_still_works() {
+    // Old callers using interrupt=true + reason should still work
+    let _g = fleet_test_guard();
+    let home = tmp_home("old-interrupt-compat");
+    std::fs::write(
+        home.join("fleet.yaml"),
+        "instances:\n  target:\n    backend: claude\n  sender:\n    backend: claude\n",
+    )
+    .ok();
+    std::env::set_var("AGEND_HOME", &home);
+    crate::tasks::handle(
+        &home,
+        "target",
+        &json!({"action": "create", "title": "busy"}),
+    );
+    let tasks = crate::tasks::handle(&home, "target", &json!({"action": "list"}));
+    let tid = tasks["tasks"][0]["id"].as_str().unwrap();
+    crate::tasks::handle(&home, "target", &json!({"action": "claim", "id": tid}));
+
+    // Use OLD names: interrupt + reason
+    let result = handle_tool(
+        "delegate_task",
+        &json!({"target_instance": "target", "task": "urgent", "interrupt": true, "reason": "legacy caller"}),
+        "sender",
+    );
+    // Should bypass busy gate (backwards-compat) + emit deprecation warning
+    assert!(
+        result.get("busy").is_none(),
+        "old interrupt=true must still bypass busy gate: {result}"
+    );
+    assert!(
+        result["warning"]
+            .as_str()
+            .unwrap_or("")
+            .contains("deprecated"),
+        "old names must emit deprecation warning: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn test_old_inbox_json_with_interrupt_meta_deserializes_into_force_meta() {
+    // Sprint 8-9 inbox JSONL uses "interrupt_meta" + "interrupted" + "interrupted_at".
+    // Must deserialize into ForceMeta via serde aliases.
+    let old_json = r#"{"schema_version":1,"id":"m-old","from":"test","text":"hi","kind":null,"timestamp":"2026-01-01T00:00:00Z","interrupt_meta":{"interrupted":true,"reason":"legacy","interrupted_at":"2026-01-01T00:00:00Z"}}"#;
+    let msg: crate::inbox::InboxMessage =
+        serde_json::from_str(old_json).expect("deserialize old format");
+    assert!(
+        msg.force_meta.is_some(),
+        "old interrupt_meta must deserialize into force_meta"
+    );
+    let meta = msg.force_meta.unwrap();
+    assert!(meta.forced, "interrupted=true must map to forced=true");
+    assert_eq!(meta.reason, "legacy");
+    assert!(!meta.forced_at.is_empty());
+}
+
+// --- Sprint 11: tool_kill pure function tests ---
+
+#[test]
+#[cfg(unix)]
+fn test_tool_kill_result_includes_pgid_and_target() {
+    let result = build_tool_kill_result("agent1", 12345, "");
+    assert_eq!(result["ok"], true);
+    assert_eq!(result["target"], "agent1");
+    assert_eq!(result["pgid"], 12345);
+    assert!(result["reason"].is_null(), "empty reason must not appear");
+}
+
+#[test]
+#[cfg(unix)]
+fn test_tool_kill_result_includes_reason_when_provided() {
+    let result = build_tool_kill_result("agent1", 12345, "stuck on cargo test");
+    assert_eq!(result["reason"], "stuck on cargo test");
+}
+
+#[test]
+#[cfg(unix)]
+fn test_tool_kill_audit_format() {
+    let audit = build_tool_kill_audit("priority task", 42);
+    assert!(audit.contains("reason=priority task"), "audit: {audit}");
+    assert!(audit.contains("pgid=42"), "audit: {audit}");
+}
+
+#[test]
+fn test_tool_kill_target_not_found_returns_error() {
+    let _g = fleet_test_guard();
+    let home = tmp_home("tool-kill-notfound");
+    std::env::set_var("AGEND_HOME", &home);
+    // No fleet.yaml → target doesn't exist
+    let result = handle_tool("tool_kill", &json!({"target": "ghost"}), "sender");
+    assert!(
+        result.get("error").is_some(),
+        "tool_kill to nonexistent target must error: {result}"
+    );
+    std::env::remove_var("AGEND_HOME");
+    std::fs::remove_dir_all(&home).ok();
+}
+
+// ─── Team auto-attach tests (Sprint 14 PR-AL) ────────────────────
+
+#[test]
+fn resolve_team_layout_auto_attaches_to_orchestrator() {
+    let home = tmp_home("team_attach");
+    let team = serde_json::json!({
+        "name": "dev", "members": ["dev-lead", "dev-impl-1"],
+        "orchestrator": "dev-lead", "created_at": "2026-01-01T00:00:00Z"
+    });
+    let store = serde_json::json!({"teams": [team]});
+    std::fs::write(home.join("teams.json"), store.to_string()).expect("write");
+    let (layout, target) = resolve_team_layout(&home, "dev-impl-1", None, None);
+    assert_eq!(layout, "split-right");
+    assert_eq!(target.as_deref(), Some("dev-lead"));
+    std::fs::remove_dir_all(&home).ok();
+}
+
+#[test]
+fn resolve_team_layout_no_team_defaults_to_tab() {
+    let home = tmp_home("no_team");
+    let (layout, target) = resolve_team_layout(&home, "standalone", None, None);
+    assert_eq!(layout, "tab");
+    assert!(target.is_none());
+}
+
+#[test]
+fn resolve_team_layout_caller_override_preserved() {
+    let home = tmp_home("override");
+    let team = serde_json::json!({
+        "name": "dev", "members": ["dev-lead", "dev-impl-1"],
+        "orchestrator": "dev-lead", "created_at": "2026-01-01T00:00:00Z"
+    });
+    let store = serde_json::json!({"teams": [team]});
+    std::fs::write(home.join("teams.json"), store.to_string()).expect("write");
+    let layout_val = serde_json::json!("tab");
+    let (layout, target) = resolve_team_layout(&home, "dev-impl-1", Some(&layout_val), None);
+    assert_eq!(
+        layout, "tab",
+        "caller explicit layout=tab must not be overridden"
+    );
+    assert!(target.is_none());
+    std::fs::remove_dir_all(&home).ok();
+}
